pyproject.toml
[tool.poetry]
name = "constellate_app"
version = "0.1.3"
description = "Render Jupyter Notebooks as beautiful interactive webapps"
authors = ["Nicholas Miklaucic <nicholas.miklaucic@gmail.com>"]
license = "MIT"
include = ["constellate-server/src/**/*", "constellate-server/public/**/*", "constellate-server/*"]
packages = [
 { include = "constellate" }
]

[tool.poetry.dependencies]
python = ">=3.8, <3.11"
yapf = "^0.32.0"
click-log = "^0.4.0"
python-slugify = "^6.1.2"
click = "^8.1.3"
watchdog = "^2.1.9"
rho-plus = "^0.4.0"
GitPython = "^3.1.27"
matplotlib = "^3.5.2"

[tool.poetry.dev-dependencies]
numpy = "^1.22.4"
pandas = "^1.4.2"
colorcet = "^3.0.0"
seaborn = "^0.11.2"
panel = "^0.13.1"
altair = "^4.2.0"

[tool.poetry.scripts]
constellate = 'constellate.cli.app:cli'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
